requirements.txt
argparse
openai
pillow
langchain
mimetypes